Fix up unit tests, I think
